+++ wezterm-mux-server-impl/src/dispatch.rs
@@ -96,6 +96,20 @@ where
                 }
                 handler.schedule_pane_push(pane_id);
             }
+            Ok(Item::Notif(MuxNotification::AssignClipboard {
+                pane_id,
+                selection,
+                clipboard,
+            })) => {
+                Pdu::SetClipboard(codec::SetClipboard {
+                    pane_id,
+                    clipboard,
+                    selection,
+                })
+                .encode_async(&mut stream, 0)
+                .await?;
+                stream.flush().await.context("flushing PDU to client")?;
+            }
             Ok(Item::Notif(MuxNotification::WindowRemoved(_window_id))) => {}
             Ok(Item::Notif(MuxNotification::WindowCreated(_window_id))) => {}
             Ok(Item::Notif(MuxNotification::WindowInvalidated(_window_id))) => {}